OBPIH-4684 Allow minimum required role assistant to delete pending putaways
* OBPIH-4684 Allow admin/manager/assistant to delete pending putaways

* OBPIH-4684 Check for minimum required role which is assistant

+++ grails-app/controllers/org/pih/warehouse/order/OrderController.groovy
@@ -300,6 +300,11 @@ class OrderController {
                 redirect(action: "show", id: orderInstance?.id)
                 return
             }
+            if (orderInstance.status != OrderStatus.PENDING) {
+                flash.message = "${warehouse.message(code: 'default.not.deleted.message', args: [warehouse.message(code: 'order.label', default: 'Order'), orderInstance.orderNumber])}"
+                redirect(action: "list", id: params.id, params: [orderType: orderInstance.orderType])
+                return
+            }
             try {
                 orderService.deleteOrder(orderInstance)
                 flash.message = "${warehouse.message(code: 'default.deleted.message', args: [warehouse.message(code: 'order.label', default: 'Order'), orderInstance.orderNumber])}"
